scripts/content-pipeline/requirements.txt
supabase>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
trafilatura>=2.0.0
google-genai>=1.0.0
pydantic>=2.0.0
